IsopycnalSurfaces

Documentation for IsopycnalSurfaces.

IsopycnalSurfaces.SeaWaterDensity

IsopycnalSurfaces.dedup!

IsopycnalSurfaces.dedupfirst!

IsopycnalSurfaces.density

IsopycnalSurfaces.density_reference_pure_water

IsopycnalSurfaces.mixinversions!

IsopycnalSurfaces.notnanorzero

IsopycnalSurfaces.secant_bulk_modulus

IsopycnalSurfaces.sigma0column

IsopycnalSurfaces.sigma1column

IsopycnalSurfaces.sigma1grid

IsopycnalSurfaces.sigmacolumn

IsopycnalSurfaces.temperature68

IsopycnalSurfaces.var2sigmacolumn

IsopycnalSurfaces.vars2sigma1

IsopycnalSurfaces.SeaWaterDensity — Function

SeaWaterDensity(Θ,Σ,Π,Π0) from MITgcmTools.jl/PhysicalOceanography.jl, From Gael Forget

Compute potential density (ρP), in situ density (ρI), and density referenced to PREF (Π0 in decibars) from potential temperature (Θ in °C), salinity (Σ in psu) and pressure (Π in decibars) according to the UNESCO / Jackett & McDougall 1994 equation of state.

Credits: code based on a Matlab implementation by B. Ferron

Reference: https://www.jodc.go.jp/info/iocdoc/UNESCOtech/059832eb.pdf

Check value: ρI = 1041.83267kg/m^3 for Θ=3°Celcius, Σ=35psu, Π=3000dbar

(ρP,ρI,ρR) = SeaWaterDensity(3.,35.5,3000.)
isapprox(ρI,1041.83267, rtol=1e-6)

IsopycnalSurfaces.dedup! — Method

function dedup!(a,b)

Remove values of a that are duplicates. Remove values of b that have the same location as the duplicates in a. The length of a and b should be identical before and after invoking this function.

Arguments

a::Vector{T}: a density variable

b::Vector{T}: an accompanying tracer variable

Arguments are mutated by the function.

IsopycnalSurfaces.dedupfirst! — Method

function dedup!(a,b)

Remove the first duplicate of a. Remove the value of b that is located at the same entry as the first duplicate in a. The length of a and b should be identical before and after invoking this function.

Arguments

a::Vector{T}: a density variable

b::Vector{T}: an accompanying tracer variable

Arguments are mutated by the function.

IsopycnalSurfaces.density — Method

density(S,T,p) and dependent functions from PhysOcean.jl/EOS80.jl, From Alexander Barth

Compute the density of sea-water (kg/m³) at the salinity S (psu, PSS-78), temperature T (degree Celsius, ITS-90) and pressure p (decibar) using the UNESCO 1983 polynomial.

Reference: Fofonoff, N.P.; Millard, R.C. (1983). Algorithms for computation of fundamental properties of seawater. UNESCO Technical Papers in Marine Science, No. 44. UNESCO: Paris. 53 pp. http://web.archive.org/web/20170103000527/http://unesdoc.unesco.org/images/0005/000598/059832eb.pdf

Check value: ρI = `1041.87651kg/m^3` for Θ=`3°Celcius`, Σ=`35.5psu`, Π=`3000dbar`
ρI = density(35.5, 3, 3000)

added by Ray, Dec 09, 2021

IsopycnalSurfaces.density_reference_pure_water — Method

density_reference_pure_water(T)

density of pure water at the temperature T (degree Celsius, ITS-90)

IsopycnalSurfaces.mixinversions! — Method

function mixinversions!(a,b)

For values of a that are not increasing, an inversion is defined. Average these values of a until values of a are sorted. Do the same averaging on the accompanying vector b.

Arguments

a::Vector{T}: a density variable

b::Vector{T}: an accompanying tracer variable

Arguments are mutated by the function.

IsopycnalSurfaces.notnanorzero — Method

function notnanorzero

 true is argument is not a NaN nor zero

IsopycnalSurfaces.secant_bulk_modulus — Method

secant_bulk_modulus(S,T,p)

Compute the secant bulk modulus of sea-water (bars) at the salinity S (psu, PSS-78), temperature T (degree Celsius, ITS-90) and pressure p (decibar) using the UNESCO polynomial 1983. Fofonoff, N.P.; Millard, R.C. (1983). Algorithms for computation of fundamental properties of seawater. UNESCO Technical Papers in Marine Science, No. 44. UNESCO: Paris. 53 pp. http://web.archive.org/web/20170103000527/http://unesdoc.unesco.org/images/0005/000598/059832eb.pdf

IsopycnalSurfaces.sigma0column — Function

function sigma0column(θ,S,p)
σ₀ for a water column
Untested for a mix of float values

Arguments

θz::Vector{T}: potential temperature

Sz::Vector{T}: practical salinity

pz::Vector{T}: vertical profile of standard pressures

Output

σ₀: sigma-0 for wet points in column

IsopycnalSurfaces.sigma1column — Function

function sigma1column(θ,S,p)
σ₁ for a water column
Untested for a mix of float values

Arguments

θz::Vector{T}: potential temperature

Sz::Vector{T}: practical salinity

pz::Vector{T}: vertical profile of standard pressures

Output

σ₁: sigma-1 for wet points in column

IsopycnalSurfaces.sigma1grid — Method

function sigma1grid() Standard (from Susan Wijffels, WHOI) choice of sigma1 surfaces

Arguments

z: value

Output

σ₁grid: list (vector) of σ₁ values

IsopycnalSurfaces.sigmacolumn — Method

function sigmacolumn(θ,S,p,p0)
σ for a water column

Arguments

θz::Vector{T}: potential temperature

Sz::Vector{T}: practical salinity

pz::Vector{T}: vertical profile of standard pressures

p0::Integer: reference pressure

Output

σ::Vector{T}: sigma for wet points in column

IsopycnalSurfaces.temperature68 — Method

temperature68(T)

Convert temperature T from ITS-90 scale to the IPTS-68 scale following Saunders, 1990. Saunders, P.M. 1990, The International Temperature Scale of 1990, ITS-90. No.10, p.10. https://web.archive.org/web/20170304194831/http://webapp1.dlib.indiana.edu/virtualdisklibrary/index.cgi/4955867/FID474/wocedocs/newsltr/news10/news10.pdf

IsopycnalSurfaces.var2sigmacolumn — Method

function var2sigmacolumn(σ,v,σ₁grid,splorder)
map θ,S, p onto σ₁ surfaces for a water column

Arguments

σ::Array{Float,1}}`: sigma values of input variable

v::Array{Float,1}}: variable of interest

sig1: σ surface values

splorder: 1-5, order of spline

linearinterp: optional argument, true to force linear interpolation

Output

θonσ: variable on sig1 sigma surfaces

IsopycnalSurfaces.vars2sigma1 — Method

function vars2sigma1(vars,p,sig1grid,γ,spline_order)
map variables from regular 3D grid onto sigma1 surfaces

Arguments

vars::Dict{String,Array{T,3}}}: dict of 3d arrays

p::Array{T,1} : vertical profile of standard pressures

sig1grid: σ₁ surface values

γ: grid description needed for preallocation

splorder: 1-5, order of spline

Output

varsσ::Dict{String,Array{T,3}: dict of 3d arrays of variables on sigma1 surfaces